64 bit hashcode works with good approximation for elements grater than 1 billion

+++ src/main/java/cardinality/hash/HashCodeProvider.java
@@ -2,6 +2,6 @@ package cardinality.hash;
 
 public interface HashCodeProvider<T> {
 
-    int hashCode(T object);
+    long hashCode(T object);
 
 }
