+++ src/main/java/cardinality/hash/SimpleHashCodeProvider.java
@@ -3,7 +3,7 @@ package cardinality.hash;
 public class SimpleHashCodeProvider<T> implements HashCodeProvider<T> {
 
     @Override
-    public int hashCode(T object) {
+    public long hashCode(T object) {
         return object.hashCode();
     }
